+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* **Tweak** - Make ACF field pitch translatable
+* **New** - ACF, secondary description for custom taxonomies
+
 ### 1.1.1: August 30th, 2018
 * **Tweak** - Change the way ACF Fields are loaded (HPP replaced JSON)
 * **New** - ACF for Package, new field pitch